airflow_setup/requirements.txt
apache-airflow-providers-postgres
apache-airflow-providers-docker
yfinance
pandas
SQLAlchemy
python-dotenv
pendulum
moexalgo
